ARM: dts: msm: Fix IOMMU address reserve definition
In order to ensure that IOMMU returns a 32 bit address, extend the
address and size cells to 2, in order to define a region that will
block 0xf0000000--0xffffffffffffffff.  By increasing the address and
size cells for the parent node, the DWC3 child node has to also
change the 'reg' entry to accommodate the added values.

Change-Id: Iaa413f0c3ac7100afea9a6d8e020bdb795866501
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>